apps/controlplane-api/requirements.txt
fastapi==0.115.0
httpx==0.27.2
PyYAML==6.0.2
uvicorn[standard]==0.30.6
